refactor: move services in different folders with their names
