+++ angular-frontend/src/app/app.module.ts
@@ -6,8 +6,8 @@ import { AppComponent } from './app.component';
 import {  ClientProfileComponent } from './pages/client-profile/client-profile.component';
 import { LoginComponent } from './pages/login/login.component';
 import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
-import { TokenInterceptor } from './service/token-interceptor.service';
-import { TipsService } from './service/tips.service';
+import { TokenInterceptor } from './service/token-interceptor/token-interceptor.service';
+import { TipsService } from './service/tips/tips.service';
 // import {  NavbarComponent } from './navbar/navbar.component';
 // import {  AccountManagementComponent } from './account-management/account-management.component';
 // import {  StatisticsComponent } from './statistics/statistics.component';
