+++ angular-frontend/src/app/service/tips/tips.service.ts
